+++ ibllib/io/video.py
@@ -30,16 +30,18 @@ def get_video_frame(video_path, frame_number):
     return frame_image
 
 
-def get_video_frames_preload(video_path, frame_numbers, mask=Ellipsis, as_list=False):
+def get_video_frames_preload(video_path, frame_numbers=None, mask=Ellipsis, as_list=False,
+                             func=lambda x: x,):
     """
     Obtain numpy array corresponding to a particular video frame in video_path.
     Fetching and returning a list is about 33% faster but may be less memory controlled. NB: Any
     gain in speed will be lost if subsequently converted to array.
     :param video_path: URL or local path to mp4 file
-    :param frame_numbers: video frames to be returned
+    :param frame_numbers: video frames to be returned. If None, return all frames.
     :param mask: a logical mask or slice to apply to frames
     :param as_list: if true the frames are returned as a list, this is faster but may be less
     memory efficient
+    :param func: Function to be applied to each frame. Applied after masking if applicable.
     :return: numpy array corresponding to frame of interest, or list if as_list is True.
     Default dimensions are (n, w, h, 3) where n = len(frame_numbers)
 
@@ -50,6 +52,9 @@ def get_video_frames_preload(video_path, frame_numbers, mask=Ellipsis, as_list=F
     cap = VideoStreamer(video_path).cap if is_url else cv2.VideoCapture(str(video_path))
     assert cap.isOpened(), 'Failed to open video'
 
+    frame_count = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
+    frame_numbers = frame_numbers if frame_numbers is not None else range(frame_count)
+
     # Setting the index is extremely slow; determine where frame index must be set
     # The first index is always explicitly set.
     to_set = np.insert(np.diff(frame_numbers), 0, 0) != 1
@@ -58,8 +63,8 @@ def get_video_frames_preload(video_path, frame_numbers, mask=Ellipsis, as_list=F
         frame_images = [None] * len(frame_numbers)
     else:
         ret, frame = cap.read()
-        frame_images = np.empty((len(frame_numbers), *frame[mask or ...].shape), np.uint8)
-
+        frame_images = np.empty((len(frame_numbers), *func(frame[mask or ...]).shape),
+                                np.uint8)
     for ii, i in enumerate(frame_numbers):
         sys.stdout.write(f'\rloading frame {ii}/{len(frame_numbers)}')
         sys.stdout.flush()
@@ -67,7 +72,7 @@ def get_video_frames_preload(video_path, frame_numbers, mask=Ellipsis, as_list=F
             cap.set(cv2.CAP_PROP_POS_FRAMES, i)
         ret, frame = cap.read()
         if ret:
-            frame_images[ii] = frame[mask]
+            frame_images[ii] = func(frame[mask or ...])
         else:
             print(f'failed to read frame #{i}')
     cap.release()